Removed FFMPEG and reformatted the entire code base

+++ android/src/com/mygdx/game/VideoHandler/VideoSplicer.java
@@ -11,7 +11,8 @@ public interface VideoSplicer {
     int getFramesProcessed();
 
     int getFrameCount();
-    long getFramesPerSecond();
+
+    float getFramesPerSecond();
 
 
     Bitmap getNextFrame(int frame);
